Updated example LoadSaveJSON
Fixed the JSON example. I left the data.json formatted as it was,
however it is not as "sleek" as it could be. The json could start with
the array straight away and that way cut away alot of the unnecessary
code required to maintain the format through load and save.