Serial: Set the library path for exported applications under Windows
With the switch to launch4j on Windows in 2.2 exported sketches that used the Serial library stopped working (see issue #2559). This hopefully fixes this by specifying java.library.path in the launcher arguments.

Note: Using %PATH% was the closest I could get to java.library.path prior to this patch, but it doesn't match 100%.