Remove "unofficial" types from printArray()
As a protest for the bureaucratic intransigence refusal to include the missing `short[]` in **printArray()**, I've made this opposite proposal!
https://github.com/processing/processing/pull/2896/files

Motive given was: "'shorts' are not supported in Processing and are omitted on purpose."
But so are both `long` & `double`; and they were contemplated into **printArray()** nonetheless!

Another case is `boolean[]`. It's more like hit 'n' miss. There are some array util methods which got it, but most don't.
Very inconsistent to say the least. So `boolean` is out for conciseness' sake!

A pity this project are kept at 7 locks. Alas...  :disappointed:
My only fear is that this pull request may end up being accepted! Oh the irony...  :weary: