Recompile jSSC binaries for Windows
Those shouldn't have functional changes compared to the binaries we used before, which were taken straight out of upstream jSSC. In preparation of implementing code for retrieving the USB properties I recompiled this on MSVC 2013, with some minimal changes to make it work. I'd be great to have this in a beta to catch any fallout (only tested it on 64-bit Windows 8.1 that had Visual Studio installed).