Workaround for JRE bug freezing the PDE
Because of a JRE bug, code completion will occasionally freeze the PDE.
Never seen this one before, however, it happened to me three times in a
row while I was giving a lecture today and I had to kill the PDE and
write all the code again. I can reproduce it back to beta 7. It is
caused by JIT compiler, workaround is to disable it for this particular
method. See https://bugs.openjdk.java.net/browse/JDK-8060036 and
http://kingsfleet.blogspot.com.br/2014/11/but-thats-impossible-or-finding-out.html