Update JNA to 4.2.1
JNA 4.2.0 incorporates a fix for Linux multi-arch systems, such as found on Raspbian (Raspberry Pi). The proposed version has an additional fix that is relevant to ARM. JNA 4.2.0 is currently used in the released processing-video, on all platforms.

With the old JNA I am getting the following error when trying to run the OpenKinect library:

java.lang.UnsatisfiedLinkError: JNA native support (com/sun/jna/linux-arm/libjnidispatch.so) not found in resource path (...)

I *could* fix this by including an updated version of JNA with the very library (that's definitely an option), but I now have the feeling that it might be benefitial to fix this and potentially other libraries by updating core instead. (Updating lib/jna.jar to the proposed version indeed makes the library work.)