looping & redraw to private, delete finished
define looping & redraw to private, because there are loop() and noLoop() function to set it. It will ensure the encapsulation of class. And in draw() function, finished should be false. Because class extends base and override draw() function, it will execute super.draw(). If "finshed" is true, it will execute once.