Scroll textarea to the left when caret is outside
Left gutter is added in Java mode,
and `0` doesn't mean the left limit for `x` position.
So Scrolling to the left is not started
even if caret is hidden with left gutter.

This change calculates leftmost position of textarea,
and when caret moves out of it,
change horizontal scroll position.