Fixed Swing high DPI scaling issues on Windows with OpenJDK 11.
As part of #5753, JEP 263 (introduced in JDK 9), offers native support for HiDPI scaling on Linux and Windows. THis causes both the JDK to naturally scale elements on HiHPI displays and then Processing again scales on its own under the assumption of Java 8 behavior, causing 2x the "correct" scale. This change migrates WindowsPlatform to simply allow the JDK to do all scaling.