No longer reference local java install for mac.
As part of #5753, one advantage to AdoptOpenJDK is that the JDK can be downloaded directly so does not require "installation" prior to use on mac. This makes mac like other OS's in that the JDK can be downloaded into the build dir instead of referencing the system wide install when building processing. This commit brings mac in line with other OS's and gets one step closer to a clean cross-os build.